config: Clear Google Maps API Key in AndroidManifest
The Google Maps API key has been removed from the `AndroidManifest.xml` file.
